# server/openapi_server/models/tag_create_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class TagCreateResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, id=None):  # noqa: E501
        """TagCreateResponse - a model defined in OpenAPI

        :param id: The id of this TagCreateResponse.  # noqa: E501
        :type id: str
        """
        self.openapi_types = {
            'id': str
        }

        self.attribute_map = {
            'id': 'id'
        }

        self._id = id

    @classmethod
    def from_dict(cls, dikt) -> 'TagCreateResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The TagCreateResponse of this TagCreateResponse.  # noqa: E501
        :rtype: TagCreateResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self):
        """Gets the id of this TagCreateResponse.

        The unique identifier of a tag  # noqa: E501

        :return: The id of this TagCreateResponse.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id):
        """Sets the id of this TagCreateResponse.

        The unique identifier of a tag  # noqa: E501

        :param id: The id of this TagCreateResponse.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id