# server/openapi_server/openapi/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: thomas.schaffter@sagebionetworks.org
    name: The ROCC Team
    url: https://Sage-Bionetworks.github.io/rocc-schemas
  description: |
    The OpenAPI specification implemented by the Challenge Registries.
    # Introduction
    TBA
  license:
    name: Apache 2.0
    url: https://Sage-Bionetworks.github.io/rocc-schemas
  title: Registry of Open Community Challenge API
  version: 0.3.0
  x-logo:
    url: https://Sage-Bionetworks.github.io/rocc-schemas/logo.png
servers:
- url: '{protocol}://rocc.org/api/v1'
  variables:
    protocol:
      default: https
      enum:
      - http
      - https
...
paths:
...
  /tags:
    delete:
      description: Delete all tags
      operationId: delete_all_tags
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete all tags
      tags:
      - Tag
      x-openapi-router-controller: openapi_server.controllers.tag_controller
    get:
      description: Returns the tags
      operationId: list_tags
      parameters:
      - description: Maximum number of results returned
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 10
          type: integer
        style: form
      - description: Index of the first result that must be returned
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfTags'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get all tags
      tags:
      - Tag
      x-openapi-router-controller: openapi_server.controllers.tag_controller
    post:
      description: Create a tag with the specified name
      operationId: create_tag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreateRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagCreateResponse'
          description: Success
          links:
            GetTagById:
              $ref: '#/components/links/GetTagById'
            DeleteTagById:
              $ref: '#/components/links/DeleteTagById'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request conflicts with current state of the target resource
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Create a tag
      tags:
      - Tag
      x-openapi-router-controller: openapi_server.controllers.tag_controller
  /tags/{tagId}:
    delete:
      description: Deletes the tag specified
      operationId: delete_tag
      parameters:
      - description: The unique identifier of the tag
        explode: false
        in: path
        name: tagId
        required: true
        schema:
          oneOf:
          - $ref: '#/components/schemas/TagId'
          - $ref: '#/components/schemas/TagName'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete a tag
      tags:
      - Tag
      x-openapi-router-controller: openapi_server.controllers.tag_controller
    get:
      description: Returns the tag specified
      operationId: get_tag
      parameters:
      - description: The unique identifier of the tag
        explode: false
        in: path
        name: tagId
        required: true
        schema:
          oneOf:
          - $ref: '#/components/schemas/TagId'
          - $ref: '#/components/schemas/TagName'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get a tag
      tags:
      - Tag
      x-openapi-router-controller: openapi_server.controllers.tag_controller
  /users:
    delete:
      description: Delete all users
      operationId: delete_all_users
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete all users
      tags:
      - User
      x-openapi-router-controller: openapi_server.controllers.user_controller
    get:
      description: Returns the users
      operationId: list_users
      parameters:
      - description: Maximum number of results returned
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 10
          type: integer
        style: form
      - description: Index of the first result that must be returned
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageOfUsers'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get all users
      tags:
      - User
      x-openapi-router-controller: openapi_server.controllers.user_controller
    post:
      description: Create a user with the specified account name
      operationId: create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
          description: Success
          links:
            GetUserById:
              $ref: '#/components/links/GetUserById'
            DeleteUserById:
              $ref: '#/components/links/DeleteUserById'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request conflicts with current state of the target resource
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Create a user
      tags:
      - User
      x-openapi-router-controller: openapi_server.controllers.user_controller
  /users/{userId}:
    delete:
      description: Deletes the user specified
      operationId: delete_user
      parameters:
      - description: The unique identifier of the user, either the user account ID
          or login
        explode: false
        in: path
        name: userId
        required: true
        schema:
          oneOf:
          - $ref: '#/components/schemas/AccountId'
          - $ref: '#/components/schemas/AccountLogin'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyObject'
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Delete a user
      tags:
      - User
      x-openapi-router-controller: openapi_server.controllers.user_controller
    get:
      description: Returns the user specified
      operationId: get_user
      parameters:
      - description: The unique identifier of the user, either the user account ID
          or login
        explode: false
        in: path
        name: userId
        required: true
        schema:
          oneOf:
          - $ref: '#/components/schemas/AccountId'
          - $ref: '#/components/schemas/AccountLogin'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: Success
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The request cannot be fulfilled due to an unexpected server
            error
      summary: Get a user
      tags:
      - User
      x-openapi-router-controller: openapi_server.controllers.user_controller
components:
  links:
...
    GetTagById:
      description: |
        The `id` value returned in the response can be used as the `tagId` parameter in `GET /tags/{tagId}`
      operationId: getTag
      parameters:
        tagId: $response.body#/tagId
    DeleteTagById:
      description: |
        The `id` value returned in the response can be used as the `tagId` parameter in `DELETE /tags/{tagId}`
      operationId: deleteTag
      parameters:
        tagId: $response.body#/tagId
    GetUserById:
      description: |
        The `userName` value returned in the response can be used as the `userName` parameter in `GET /users/{userName}`
      operationId: getUser
      parameters:
        userName: $response.body#/userName
    DeleteUserById:
      description: |
        The `userName` value returned in the response can be used as the `userName` parameter in `DELETE /users/{userName}`
      operationId: deleteUser
      parameters:
        userName: $response.body#/userName
...
  schemas:
    TagId:
      description: The unique identifier of a tag
      example: 507f1f77bcf86cd799439011
      title: TagId
      type: string
...
    ResponsePageMetadata:
      description: A page of results
      properties:
        offset:
          description: Index of the first result that must be returned
          type: integer
        limit:
          description: Maximum number of results returned
          type: integer
        paging:
          $ref: '#/components/schemas/ResponsePageMetadata_paging'
        totalResults:
          description: Total number of results in the result set
          type: integer
      required:
      - limit
      - offset
      - paging
      type: object
...
    AccountId:
      description: The unique identifier of an account
      example: 507f1f77bcf86cd799439011
      title: AccountId
      type: string
...
    Error:
      description: Problem details (tools.ietf.org/html/rfc7807)
      properties:
        title:
          description: A human readable documentation for the problem type
          title: title
          type: string
        status:
          description: The HTTP status code
          title: status
          type: integer
        detail:
          description: A human readable explanation specific to this occurrence of
            the problem
          title: detail
          type: string
        type:
          description: An absolute URI that identifies the problem type
          title: type
          type: string
      required:
      - status
      - title
      title: Error
      type: object
    EmptyObject:
      description: Empty JSON object
      type: object
...
    AccountLogin:
      description: The user or organization account name
      example: awesome-user
      maxLength: 25
      minLength: 3
      pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
      type: string
...
    TagCreateResponse:
      description: The unique identifier of the tag created
      example:
        id: 507f1f77bcf86cd799439011
      properties:
        id:
          description: The unique identifier of a tag
          example: 507f1f77bcf86cd799439011
          title: TagId
          type: string
      required:
      - id
      title: TagCreateResponse
      type: object
    TagName:
      description: The name of a tag
      example: awesome-tag
      maxLength: 30
      minLength: 3
      pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
      type: string
    TagCreateRequest:
      description: The information required to create a tag
      example:
        name: awesome-tag
      properties:
        name:
          description: The name of a tag
          example: awesome-tag
          maxLength: 30
          minLength: 3
          pattern: ^[a-z0-9]+(?:-[a-z0-9]+)*$
          type: string
      required:
      - name
      type: object
    Tag:
      allOf:
      - $ref: '#/components/schemas/TagCreateResponse'
      - $ref: '#/components/schemas/TagCreateRequest'
      description: A tag that can be assigned to a resource
      example:
        id: 507f1f77bcf86cd799439011
        name: awesome-tag
      required:
      - id
      - name
      title: Tag
      type: object
    PageOfTags:
      allOf:
      - $ref: '#/components/schemas/ResponsePageMetadata'
      - $ref: '#/components/schemas/PageOfTags_allOf'
      description: A page of Tags
      title: PageOfTags
      type: object
    UserCreateResponse:
      description: The unique identifier of the user created
      example:
        id: 507f1f77bcf86cd799439011
      properties:
        id:
          description: The unique identifier of an account
          example: 507f1f77bcf86cd799439011
          title: AccountId
          type: string
      required:
      - id
      title: UserCreateResponse
      type: object
    UserCreateRequest:
      description: The information required to create a user account
      example:
        login: awesome-user
        email: awesome-user@example.org
      properties:
        login:
          type: string
        email:
          description: An email address
          example: john.smith@example.com
          format: email
          type: string
      required:
      - email
      - login
      type: object
    User:
      allOf:
      - $ref: '#/components/schemas/UserCreateResponse'
      - $ref: '#/components/schemas/UserCreateRequest'
      - $ref: '#/components/schemas/User_allOf'
      description: A simple user
      example:
        id: 507f1f77bcf86cd799439011
        login: awesome-user
        email: awesome-user@example.org
        name: Awesome User
        avatarUrl: https://example.com/awesome-avatar.png
        createdAt: 2017-07-08T16:18:44-04:00
        updatedAt: 2017-07-08T16:18:44-04:00
        type: User
      required:
      - avatarUrl
      - createdAt
      - email
      - id
      - login
      - type
      - updatedAt
      title: User
      type: object
    PageOfUsers:
      allOf:
      - $ref: '#/components/schemas/ResponsePageMetadata'
      - $ref: '#/components/schemas/PageOfUsers_allOf'
      description: A page of Users
      title: PageOfUsers
      type: object
...
    ResponsePageMetadata_paging:
      description: Links to navigate to different pages of results
      properties:
        next:
          description: Link to the next page of results
          format: uri
          title: next
          type: string
      title: ResponsePageMetadata_paging
      type: object
...
    PageOfTags_allOf:
      properties:
        tags:
          description: An array of Tags
          items:
            $ref: '#/components/schemas/Tag'
          type: array
      required:
      - tags
      type: object
    User_allOf:
      properties:
        name:
          nullable: true
          type: string
        avatarUrl:
          example: https://example.com/awesome-avatar.png
          format: uri
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        type:
          example: User
          type: string
      type: object
    PageOfUsers_allOf:
      properties:
        users:
          description: An array of Users
          items:
            $ref: '#/components/schemas/User'
          type: array
      required:
      - users
      type: object